refactor(ios): shrink oversized CloudKit sync services (#346)
Pure restructuring of the highest-risk group, behavior-preserving; no
suppressions. spec/services/cloudkit-sync.md verified unchanged in meaning
(it names symbols and behavior, not file layout).

- CKRecordMapper (876 lines) split per record type, continuing the
  +SetMeasurement pattern: +Gyms, +Plans, +Sessions, +Settings,
  +ActiveSession; mergeIncoming/createCKRecord/serverRecordIsNewer routed
  through per-tier helpers with identical decision order. Schema gate: all
  65 record[...] key lines reappear character-identical;
  tools/check_ckrecord_drift.py clean.
- CKSyncEngineManager (801 lines) split into +Recovery and +SyncEvents;
  handleFetchedChanges (c14, 98 lines) decomposed into sorted-merge/
  dependency-retry/deletion helpers with verbatim case bodies; handleEvent
  split by event family with @unknown default behavior preserved.
- OutboxPusherService push pipeline moved to a private extension with
  pushOne/classifyPushFailure decomposed (decision order unchanged);
  InboxPollerService poll loop split into fetch/download/prune;
  SyncSessionGuard's validateAndRestore extracted inside the same write
  transaction (single-transaction guarantee holds); CKSyncConflictResolver's
  handleSentChanges split per failure class (CKError branch order, log
  strings, Sentry keys unchanged).

SwiftLint now reports ZERO violations repo-wide; baseline regenerated to an
empty ledger (21 -> 0, removals only). Sync suites: 83 + 75 tests pass;
full unit suite: 914 tests, 0 failures.

Part of #351. Closes #346.
