+++ mobile-apps/ios/LiftMark/Services/CKRecordMapper.swift
@@ -4,6 +4,14 @@ import GRDB
 /// Handles conversion between GRDB row types and CKRecords, plus merging incoming
 /// CKRecords into the local database. Extracted from CloudKitService to support the
 /// CKSyncEngine migration.
+///
+/// Per-record-type mapping and merging lives in extensions:
+/// - `CKRecordMapper+Gyms` — Gym, GymEquipment
+/// - `CKRecordMapper+Plans` — WorkoutPlan, PlannedExercise, PlannedSet
+/// - `CKRecordMapper+Sessions` — WorkoutSession, SessionExercise, SessionSet
+/// - `CKRecordMapper+Settings` — UserSettings
+/// - `CKRecordMapper+SetMeasurement` — SetMeasurement (plus set-measurement dual read/write)
+/// - `CKRecordMapper+ActiveSession` — active-session sync protection
 final class CKRecordMapper {
     let dbManager: DatabaseManager
 
@@ -77,7 +85,7 @@ final class CKRecordMapper {
         record[key] as? Date
     }
 
-    private func stringListField(_ record: CKRecord, _ key: String) -> [String] {
+    func stringListField(_ record: CKRecord, _ key: String) -> [String] {
         record[key] as? [String] ?? []
     }
 
@@ -97,36 +105,10 @@ final class CKRecordMapper {
         do {
             let dbQueue = try dbManager.database()
             return try dbQueue.read { db in
-                let recordName = record.recordID.recordName
                 let remoteDate = self.dateField(record, "updatedAt")
 
                 // Look up the local updatedAt based on record type
-                let localUpdatedAt: String? = try {
-                    switch record.recordType {
-                    case "Gym":
-                        return try GymRow.fetchOne(db, key: recordName)?.updatedAt
-                    case "GymEquipment":
-                        return try GymEquipmentRow.fetchOne(db, key: recordName)?.updatedAt
-                    case "WorkoutPlan":
-                        return try WorkoutPlanRow.fetchOne(db, key: recordName)?.updatedAt
-                    case "PlannedExercise":
-                        return try PlannedExerciseRow.fetchOne(db, key: recordName)?.updatedAt
-                    case "PlannedSet":
-                        return try PlannedSetRow.fetchOne(db, key: recordName)?.updatedAt
-                    case "WorkoutSession":
-                        return try WorkoutSessionRow.fetchOne(db, key: recordName)?.updatedAt
-                    case "SessionExercise":
-                        return try SessionExerciseRow.fetchOne(db, key: recordName)?.updatedAt
-                    case "SessionSet":
-                        return try SessionSetRow.fetchOne(db, key: recordName)?.updatedAt
-                    case "UserSettings":
-                        return try UserSettingsRow.fetchOne(db, key: recordName)?.updatedAt
-                    case "SetMeasurement":
-                        return try SetMeasurementRow.fetchOne(db, key: recordName)?.updatedAt
-                    default:
-                        return nil
-                    }
-                }()
+                let localUpdatedAt = try self.localUpdatedAt(for: record, db: db)
 
                 return self.remoteIsNewer(remoteDate: remoteDate, localUpdatedAt: localUpdatedAt)
             }
@@ -134,6 +116,43 @@ final class CKRecordMapper {
             // If we can't read the local DB, default to server wins
             Logger.shared.error(.sync, "Failed to read local record for conflict check: \(error.localizedDescription)")
             return true
+        }
+    }
+
+    /// Look up the local row's updatedAt for the given record's type and record name.
+    private func localUpdatedAt(for record: CKRecord, db: Database) throws -> String? {
+        let recordName = record.recordID.recordName
+        switch record.recordType {
+        case "Gym":
+            return try GymRow.fetchOne(db, key: recordName)?.updatedAt
+        case "GymEquipment":
+            return try GymEquipmentRow.fetchOne(db, key: recordName)?.updatedAt
+        case "WorkoutPlan":
+            return try WorkoutPlanRow.fetchOne(db, key: recordName)?.updatedAt
+        case "PlannedExercise":
+            return try PlannedExerciseRow.fetchOne(db, key: recordName)?.updatedAt
+        case "PlannedSet":
+            return try PlannedSetRow.fetchOne(db, key: recordName)?.updatedAt
+        default:
+            return try sessionTierUpdatedAt(recordType: record.recordType, recordName: recordName, db: db)
+        }
+    }
+
+    /// Session-tier and singleton half of `localUpdatedAt(for:db:)`.
+    private func sessionTierUpdatedAt(recordType: String, recordName: String, db: Database) throws -> String? {
+        switch recordType {
+        case "WorkoutSession":
+            return try WorkoutSessionRow.fetchOne(db, key: recordName)?.updatedAt
+        case "SessionExercise":
+            return try SessionExerciseRow.fetchOne(db, key: recordName)?.updatedAt
+        case "SessionSet":
+            return try SessionSetRow.fetchOne(db, key: recordName)?.updatedAt
+        case "UserSettings":
+            return try UserSettingsRow.fetchOne(db, key: recordName)?.updatedAt
+        case "SetMeasurement":
+            return try SetMeasurementRow.fetchOne(db, key: recordName)?.updatedAt
+        default:
+            return nil
         }
     }
 
@@ -151,120 +170,6 @@ final class CKRecordMapper {
         return CKRecord.Reference(recordID: id, action: .none)
     }
 
-    // MARK: - To CKRecord
-
-    func toCKRecord(_ gym: GymRow, zoneID: CKRecordZone.ID) -> CKRecord {
-        let record = CKRecord(recordType: "Gym", recordID: CKRecord.ID(recordName: gym.id, zoneID: zoneID))
-        record["name"] = gym.name as CKRecordValue
-        record["isDefault"] = Int64(gym.isDefault) as CKRecordValue
-        if let date = parseDate(gym.createdAt) { record["createdAt"] = date as CKRecordValue }
-        if let date = parseDate(gym.updatedAt) { record["updatedAt"] = date as CKRecordValue }
-        return record
-    }
-
-    func toCKRecord(_ eq: GymEquipmentRow, zoneID: CKRecordZone.ID) -> CKRecord {
-        let record = CKRecord(recordType: "GymEquipment", recordID: CKRecord.ID(recordName: eq.id, zoneID: zoneID))
-        record["name"] = eq.name as CKRecordValue
-        record["isAvailable"] = Int64(eq.isAvailable) as CKRecordValue
-        if let gymId = eq.gymId {
-            record["gymId"] = makeReference(recordName: gymId, zoneID: zoneID) as CKRecordValue
-        }
-        if let date = parseDate(eq.lastCheckedAt) { record["lastCheckedAt"] = date as CKRecordValue }
-        if let date = parseDate(eq.createdAt) { record["createdAt"] = date as CKRecordValue }
-        if let date = parseDate(eq.updatedAt) { record["updatedAt"] = date as CKRecordValue }
-        return record
-    }
-
-    func toCKRecord(_ plan: WorkoutPlanRow, zoneID: CKRecordZone.ID) -> CKRecord {
-        let record = CKRecord(recordType: "WorkoutPlan", recordID: CKRecord.ID(recordName: plan.id, zoneID: zoneID))
-        record["name"] = plan.name as CKRecordValue
-        record["isFavorite"] = Int64(plan.isFavorite) as CKRecordValue
-        if let description = plan.description { record["planDescription"] = description as CKRecordValue }
-        if let tags = plan.tags { record["tags"] = tags as CKRecordValue }
-        if let unit = plan.defaultWeightUnit { record["defaultWeightUnit"] = unit as CKRecordValue }
-        if let markdown = plan.sourceMarkdown { record["sourceMarkdown"] = markdown as CKRecordValue }
-        if let date = parseDate(plan.createdAt) { record["createdAt"] = date as CKRecordValue }
-        if let date = parseDate(plan.updatedAt) { record["updatedAt"] = date as CKRecordValue }
-        return record
-    }
-
-    func toCKRecord(_ ex: PlannedExerciseRow, zoneID: CKRecordZone.ID) -> CKRecord {
-        let record = CKRecord(recordType: "PlannedExercise", recordID: CKRecord.ID(recordName: ex.id, zoneID: zoneID))
-        record["workoutPlanId"] = makeReference(recordName: ex.workoutTemplateId, zoneID: zoneID) as CKRecordValue
-        record["exerciseName"] = ex.exerciseName as CKRecordValue
-        record["orderIndex"] = Int64(ex.orderIndex) as CKRecordValue
-        if let notes = ex.notes { record["notes"] = notes as CKRecordValue }
-        if let equipmentType = ex.equipmentType { record["equipmentType"] = equipmentType as CKRecordValue }
-        if let groupType = ex.groupType { record["groupType"] = groupType as CKRecordValue }
-        if let groupName = ex.groupName { record["groupName"] = groupName as CKRecordValue }
-        if let parentId = ex.parentExerciseId {
-            record["parentExerciseId"] = makeReference(recordName: parentId, zoneID: zoneID) as CKRecordValue
-        }
-        if let date = parseDate(ex.updatedAt) { record["updatedAt"] = date as CKRecordValue }
-        return record
-    }
-
-    func toCKRecord(_ session: WorkoutSessionRow, zoneID: CKRecordZone.ID) -> CKRecord {
-        let record = CKRecord(recordType: "WorkoutSession", recordID: CKRecord.ID(recordName: session.id, zoneID: zoneID))
-        record["name"] = session.name as CKRecordValue
-        record["date"] = session.date as CKRecordValue
-        record["status"] = session.status as CKRecordValue
-        if let pid = session.workoutTemplateId { record["workoutPlanId"] = pid as CKRecordValue }
-        if let date = parseDate(session.startTime) { record["startTime"] = date as CKRecordValue }
-        if let date = parseDate(session.endTime) { record["endTime"] = date as CKRecordValue }
-        if let dur = session.duration { record["duration"] = Int64(dur) as CKRecordValue }
-        if let notes = session.notes { record["notes"] = notes as CKRecordValue }
-        if let date = parseDate(session.updatedAt) { record["updatedAt"] = date as CKRecordValue }
-        return record
-    }
-
-    func toCKRecord(_ se: SessionExerciseRow, zoneID: CKRecordZone.ID) -> CKRecord {
-        let record = CKRecord(recordType: "SessionExercise", recordID: CKRecord.ID(recordName: se.id, zoneID: zoneID))
-        record["workoutSessionId"] = makeReference(recordName: se.workoutSessionId, zoneID: zoneID) as CKRecordValue
-        record["exerciseName"] = se.exerciseName as CKRecordValue
-        record["orderIndex"] = Int64(se.orderIndex) as CKRecordValue
-        record["status"] = se.status as CKRecordValue
-        if let notes = se.notes { record["notes"] = notes as CKRecordValue }
-        if let equipmentType = se.equipmentType { record["equipmentType"] = equipmentType as CKRecordValue }
-        if let groupType = se.groupType { record["groupType"] = groupType as CKRecordValue }
-        if let groupName = se.groupName { record["groupName"] = groupName as CKRecordValue }
-        if let parentId = se.parentExerciseId {
-            record["parentExerciseId"] = makeReference(recordName: parentId, zoneID: zoneID) as CKRecordValue
-        }
-        if let date = parseDate(se.updatedAt) { record["updatedAt"] = date as CKRecordValue }
-        return record
-    }
-
-    func toCKRecord(_ settings: UserSettingsRow, zoneID: CKRecordZone.ID) -> CKRecord {
-        let record = CKRecord(
-            recordType: "UserSettings",
-            recordID: CKRecord.ID(recordName: "user-settings", zoneID: zoneID)
-        )
-        record["defaultWeightUnit"] = settings.defaultWeightUnit as CKRecordValue
-        record["enableWorkoutTimer"] = Int64(settings.enableWorkoutTimer) as CKRecordValue
-        record["autoStartRestTimer"] = Int64(settings.autoStartRestTimer) as CKRecordValue
-        record["theme"] = settings.theme as CKRecordValue
-        record["notificationsEnabled"] = Int64(settings.notificationsEnabled) as CKRecordValue
-        record["healthKitEnabled"] = Int64(settings.healthkitEnabled) as CKRecordValue
-        record["liveActivitiesEnabled"] = Int64(settings.liveActivitiesEnabled) as CKRecordValue
-        record["keepScreenAwake"] = Int64(settings.keepScreenAwake) as CKRecordValue
-        record["showOpenInClaudeButton"] = Int64(settings.showOpenInClaudeButton) as CKRecordValue
-        record["countdownSoundsEnabled"] = Int64(settings.countdownSoundsEnabled) as CKRecordValue
-        record["defaultTimerCountdown"] = Int64(settings.defaultTimerCountdown) as CKRecordValue
-        record["defaultWeightStepLbs"] = settings.defaultWeightStepLbs as CKRecordValue
-        record["aiPromptIncludeFormatPointer"] = Int64(settings.aiPromptIncludeFormatPointer) as CKRecordValue
-        record["aiPromptIncludeRecentWorkouts"] = Int64(settings.aiPromptIncludeRecentWorkouts) as CKRecordValue
-        record["aiPromptIncludeProgression"] = Int64(settings.aiPromptIncludeProgression) as CKRecordValue
-        record["aiPromptIncludeEquipment"] = Int64(settings.aiPromptIncludeEquipment) as CKRecordValue
-        if let addition = settings.customPromptAddition {
-            record["customPromptAddition"] = addition as CKRecordValue
-        }
-        if let tiles = settings.homeTiles { record["homeTiles"] = tiles as CKRecordValue }
-        if let date = parseDate(settings.updatedAt) { record["updatedAt"] = date as CKRecordValue }
-        // Never sync anthropicApiKey
-        return record
-    }
-
     // MARK: - Merge Incoming
 
     /// Routes an incoming CKRecord to the appropriate merge method. Returns true if local DB was updated.
@@ -279,424 +184,25 @@ final class CKRecordMapper {
         metadataStore.save(record)
 
         switch record.recordType {
-        case "Gym":
-            return try mergeGym(record, dbQueue: dbQueue)
-        case "GymEquipment":
-            return try mergeGymEquipment(record, dbQueue: dbQueue)
-        case "WorkoutPlan":
-            return try mergeWorkoutPlan(record, dbQueue: dbQueue)
-        case "PlannedExercise":
-            return try mergePlannedExercise(record, dbQueue: dbQueue)
-        case "PlannedSet":
-            return try mergePlannedSet(record, dbQueue: dbQueue)
-        case "WorkoutSession":
-            return try mergeWorkoutSession(record, dbQueue: dbQueue)
-        case "SessionExercise":
-            return try mergeSessionExercise(record, dbQueue: dbQueue)
-        case "SessionSet":
-            return try mergeSessionSet(record, dbQueue: dbQueue)
+        case "Gym", "GymEquipment":
+            return try mergeGymRecord(record, dbQueue: dbQueue)
+        case "WorkoutPlan", "PlannedExercise", "PlannedSet":
+            return try mergePlanRecord(record, dbQueue: dbQueue)
+        case "WorkoutSession", "SessionExercise", "SessionSet":
+            return try mergeSessionRecord(record, dbQueue: dbQueue)
         case "UserSettings":
             return try mergeUserSettings(record, dbQueue: dbQueue)
         case "SetMeasurement":
             return try mergeSetMeasurement(record, dbQueue: dbQueue)
         default:
-            Logger.shared.warn(.sync, "Unknown record type for merge: \(record.recordType)")
-            return false
+            return logUnknownRecordType(record)
         }
     }
 
-    // MARK: - Individual Merge Methods
-
-    private func mergeGym(_ record: CKRecord, dbQueue: DatabaseQueue) throws -> Bool {
-        let remoteUpdatedAt = dateField(record, "updatedAt")
-        return try dbQueue.write { db in
-            let existing = try GymRow.fetchOne(db, key: record.recordID.recordName)
-
-            // Don't re-insert a gym that was soft-deleted locally
-            if let existing, existing.deletedAt != nil {
-                return false
-            }
-
-            if let existing, !self.remoteIsNewer(remoteDate: remoteUpdatedAt, localUpdatedAt: existing.updatedAt) {
-                return false
-            }
-            let row = GymRow(
-                id: record.recordID.recordName,
-                name: self.stringField(record, "name") ?? "Gym",
-                isDefault: Int(self.int64Field(record, "isDefault") ?? 0),
-                deletedAt: nil,
-                createdAt: self.dateToISO(self.dateField(record, "createdAt")) ?? existing?.createdAt
-                    ?? self.isoFormatter.string(from: Date()),
-                updatedAt: self.dateToISO(remoteUpdatedAt) ?? existing?.updatedAt
-                    ?? self.isoFormatter.string(from: Date())
-            )
-            if existing != nil { try row.update(db) } else { try row.insert(db) }
-            return true
-        }
-    }
-
-    private func mergeGymEquipment(_ record: CKRecord, dbQueue: DatabaseQueue) throws -> Bool {
-        let remoteUpdatedAt = dateField(record, "updatedAt")
-        return try dbQueue.write { db in
-            let existing = try GymEquipmentRow.fetchOne(db, key: record.recordID.recordName)
-
-            // Don't re-insert equipment that was soft-deleted locally
-            if let existing, existing.deletedAt != nil {
-                return false
-            }
-
-            if let existing, !self.remoteIsNewer(remoteDate: remoteUpdatedAt, localUpdatedAt: existing.updatedAt) {
-                return false
-            }
-            let row = GymEquipmentRow(
-                id: record.recordID.recordName,
-                name: self.stringField(record, "name") ?? "Equipment",
-                isAvailable: Int(self.int64Field(record, "isAvailable") ?? 1),
-                lastCheckedAt: self.dateToISO(self.dateField(record, "lastCheckedAt")),
-                deletedAt: nil,
-                createdAt: self.dateToISO(self.dateField(record, "createdAt")) ?? existing?.createdAt
-                    ?? self.isoFormatter.string(from: Date()),
-                updatedAt: self.dateToISO(remoteUpdatedAt) ?? existing?.updatedAt
-                    ?? self.isoFormatter.string(from: Date()),
-                gymId: self.referenceId(record, "gymId")
-            )
-            if existing != nil { try row.update(db) } else { try row.insert(db) }
-            return true
-        }
-    }
-
-    private func mergeWorkoutPlan(_ record: CKRecord, dbQueue: DatabaseQueue) throws -> Bool {
-        let remoteUpdatedAt = dateField(record, "updatedAt")
-        return try dbQueue.write { db in
-            let existing = try WorkoutPlanRow.fetchOne(db, key: record.recordID.recordName)
-            if let existing, !self.remoteIsNewer(remoteDate: remoteUpdatedAt, localUpdatedAt: existing.updatedAt) {
-                return false
-            }
-            let row = WorkoutPlanRow(
-                id: record.recordID.recordName,
-                name: self.stringField(record, "name") ?? "Workout",
-                description: self.stringField(record, "planDescription"),
-                tags: self.stringField(record, "tags"),
-                defaultWeightUnit: self.stringField(record, "defaultWeightUnit"),
-                sourceMarkdown: self.stringField(record, "sourceMarkdown"),
-                createdAt: self.dateToISO(self.dateField(record, "createdAt")) ?? existing?.createdAt
-                    ?? self.isoFormatter.string(from: Date()),
-                updatedAt: self.dateToISO(remoteUpdatedAt) ?? existing?.updatedAt
-                    ?? self.isoFormatter.string(from: Date()),
-                isFavorite: Int(self.int64Field(record, "isFavorite") ?? 0)
-            )
-            if existing != nil { try row.update(db) } else { try row.insert(db) }
-            return true
-        }
-    }
-
-    private func mergePlannedExercise(_ record: CKRecord, dbQueue: DatabaseQueue) throws -> Bool {
-        let remoteUpdatedAt = dateField(record, "updatedAt")
-        return try dbQueue.write { db in
-            let existing = try PlannedExerciseRow.fetchOne(db, key: record.recordID.recordName)
-            let fk = self.referenceId(record, "workoutPlanId") ?? existing?.workoutTemplateId ?? ""
-            if fk.isEmpty && existing == nil {
-                Logger.shared.error(
-                    .sync,
-                    "[sync-merge] Skipping PlannedExercise \(record.recordID.recordName): missing workoutPlanId FK"
-                )
-                return false
-            }
-
-            if let existing, !self.remoteIsNewer(remoteDate: remoteUpdatedAt, localUpdatedAt: existing.updatedAt) {
-                return false
-            }
-
-            let row = PlannedExerciseRow(
-                id: record.recordID.recordName,
-                workoutTemplateId: fk,
-                exerciseName: self.stringField(record, "exerciseName") ?? existing?.exerciseName ?? "",
-                orderIndex: Int(self.int64Field(record, "orderIndex") ?? Int64(existing?.orderIndex ?? 0)),
-                notes: self.stringField(record, "notes"),
-                equipmentType: self.stringField(record, "equipmentType"),
-                groupType: self.stringField(record, "groupType"),
-                groupName: self.stringField(record, "groupName"),
-                parentExerciseId: self.referenceId(record, "parentExerciseId"),
-                updatedAt: self.dateToISO(remoteUpdatedAt) ?? existing?.updatedAt
-            )
-            if existing != nil {
-                try row.update(db)
-            } else {
-                let duplicate = try PlannedExerciseRow
-                    .filter(Column("workout_template_id") == row.workoutTemplateId)
-                    .filter(Column("exercise_name") == row.exerciseName)
-                    .filter(Column("order_index") == row.orderIndex)
-                    .fetchOne(db)
-                if duplicate != nil {
-                    Logger.shared.warn(
-                        .sync,
-                        "Skipping duplicate exercise: \(row.exerciseName) at index \(row.orderIndex)"
-                    )
-                    return false
-                }
-                try row.insert(db)
-            }
-            return true
-        }
-    }
-
-    private func mergePlannedSet(_ record: CKRecord, dbQueue: DatabaseQueue) throws -> Bool {
-        let remoteUpdatedAt = dateField(record, "updatedAt")
-        return try dbQueue.write { db in
-            let setId = record.recordID.recordName
-            let existing = try PlannedSetRow.fetchOne(db, key: setId)
-            let fk = self.referenceId(record, "plannedExerciseId") ?? existing?.templateExerciseId ?? ""
-            if fk.isEmpty && existing == nil {
-                Logger.shared.error(.sync, "[sync-merge] Skipping PlannedSet \(setId): missing plannedExerciseId FK")
-                return false
-            }
-
-            if let existing, !self.remoteIsNewer(remoteDate: remoteUpdatedAt, localUpdatedAt: existing.updatedAt) {
-                return false
-            }
-
-            let attrs = self.stringListField(record, "attributes")
-            let now = self.dateToISO(remoteUpdatedAt) ?? existing?.updatedAt
-            let row = PlannedSetRow(
-                id: setId,
-                templateExerciseId: fk,
-                orderIndex: Int(self.int64Field(record, "orderIndex") ?? Int64(existing?.orderIndex ?? 0)),
-                restSeconds: self.int64Field(record, "restSeconds").map { Int($0) },
-                isDropset: attrs.contains("dropset") ? 1 : 0,
-                isPerSide: attrs.contains("perSide") ? 1 : 0,
-                isAmrap: attrs.contains("amrap") ? 1 : 0,
-                notes: self.stringField(record, "notes"),
-                updatedAt: now
-            )
-            if existing != nil {
-                try row.update(db)
-            } else {
-                let duplicate = try PlannedSetRow
-                    .filter(Column("template_exercise_id") == row.templateExerciseId)
-                    .filter(Column("order_index") == row.orderIndex)
-                    .fetchOne(db)
-                if duplicate != nil {
-                    Logger.shared.warn(.sync, "Skipping duplicate set at index \(row.orderIndex)")
-                    return false
-                }
-                try row.insert(db)
-            }
-
-            // Replace measurements from CK record fields (dual-read: old-format CKRecords)
-            try db.execute(
-                sql: "DELETE FROM set_measurements WHERE set_id = ? AND parent_type = 'planned'",
-                arguments: [setId]
-            )
-            try self.insertMeasurementsFromCKRecord(
-                record, setId: setId, parentType: "planned", role: "target", now: now, in: db
-            )
-
-            return true
-        }
-    }
-
-    private func mergeWorkoutSession(_ record: CKRecord, dbQueue: DatabaseQueue) throws -> Bool {
-        let remoteUpdatedAt = dateField(record, "updatedAt")
-        return try dbQueue.write { db in
-            let existing = try WorkoutSessionRow.fetchOne(db, key: record.recordID.recordName)
-
-            if let existing, !self.remoteIsNewer(remoteDate: remoteUpdatedAt, localUpdatedAt: existing.updatedAt) {
-                return false
-            }
-
-            // Don't let remote data overwrite a local cancellation
-            let remoteStatus = self.stringField(record, "status")
-            let mergedStatus: String
-            if existing?.status == SessionStatus.canceled.rawValue {
-                mergedStatus = SessionStatus.canceled.rawValue
-            } else {
-                mergedStatus = remoteStatus ?? existing?.status ?? SessionStatus.inProgress.rawValue
-            }
-
-            let row = WorkoutSessionRow(
-                id: record.recordID.recordName,
-                workoutTemplateId: self.stringField(record, "workoutPlanId"),
-                name: self.stringField(record, "name") ?? existing?.name ?? "Workout",
-                date: self.stringField(record, "date") ?? existing?.date ?? "",
-                startTime: self.dateToISO(self.dateField(record, "startTime")),
-                endTime: self.dateToISO(self.dateField(record, "endTime")),
-                duration: self.int64Field(record, "duration").map { Int($0) },
-                notes: self.stringField(record, "notes"),
-                status: mergedStatus,
-                updatedAt: self.dateToISO(remoteUpdatedAt) ?? existing?.updatedAt
-            )
-            if existing != nil { try row.update(db) } else { try row.insert(db) }
-            return true
-        }
-    }
-
-    private func mergeSessionExercise(_ record: CKRecord, dbQueue: DatabaseQueue) throws -> Bool {
-        let remoteUpdatedAt = dateField(record, "updatedAt")
-        return try dbQueue.write { db in
-            let existing = try SessionExerciseRow.fetchOne(db, key: record.recordID.recordName)
-            let fk = self.referenceId(record, "workoutSessionId") ?? existing?.workoutSessionId ?? ""
-            if fk.isEmpty && existing == nil {
-                Logger.shared.error(
-                    .sync,
-                    "[sync-merge] Skipping SessionExercise \(record.recordID.recordName): missing workoutSessionId FK"
-                )
-                return false
-            }
-
-            if let existing, !self.remoteIsNewer(remoteDate: remoteUpdatedAt, localUpdatedAt: existing.updatedAt) {
-                return false
-            }
-
-            let row = SessionExerciseRow(
-                id: record.recordID.recordName,
-                workoutSessionId: fk,
-                exerciseName: self.stringField(record, "exerciseName") ?? existing?.exerciseName ?? "",
-                orderIndex: Int(self.int64Field(record, "orderIndex") ?? Int64(existing?.orderIndex ?? 0)),
-                notes: self.stringField(record, "notes"),
-                equipmentType: self.stringField(record, "equipmentType"),
-                groupType: self.stringField(record, "groupType"),
-                groupName: self.stringField(record, "groupName"),
-                parentExerciseId: self.referenceId(record, "parentExerciseId"),
-                status: self.stringField(record, "status") ?? existing?.status ?? ExerciseStatus.pending.rawValue,
-                updatedAt: self.dateToISO(remoteUpdatedAt) ?? existing?.updatedAt
-            )
-            if existing != nil { try row.update(db) } else { try row.insert(db) }
-            return true
-        }
-    }
-
-    private func mergeSessionSet(_ record: CKRecord, dbQueue: DatabaseQueue) throws -> Bool {
-        let remoteUpdatedAt = dateField(record, "updatedAt")
-        return try dbQueue.write { db in
-            let setId = record.recordID.recordName
-            let existing = try SessionSetRow.fetchOne(db, key: setId)
-            let fk = self.referenceId(record, "sessionExerciseId") ?? existing?.sessionExerciseId ?? ""
-            if fk.isEmpty && existing == nil {
-                Logger.shared.error(.sync, "[sync-merge] Skipping SessionSet \(setId): missing sessionExerciseId FK")
-                return false
-            }
-
-            if let existing, !self.remoteIsNewer(remoteDate: remoteUpdatedAt, localUpdatedAt: existing.updatedAt) {
-                return false
-            }
-
-            let attrs = self.stringListField(record, "attributes")
-            let now = self.dateToISO(remoteUpdatedAt) ?? existing?.updatedAt
-            let row = SessionSetRow(
-                id: setId,
-                sessionExerciseId: fk,
-                orderIndex: Int(self.int64Field(record, "orderIndex") ?? Int64(existing?.orderIndex ?? 0)),
-                restSeconds: self.int64Field(record, "restSeconds").map { Int($0) },
-                completedAt: self.dateToISO(self.dateField(record, "completedAt")),
-                status: self.stringField(record, "status") ?? existing?.status ?? SetStatus.pending.rawValue,
-                notes: self.stringField(record, "notes"),
-                isDropset: attrs.contains("dropset") ? 1 : 0,
-                isPerSide: attrs.contains("perSide") ? 1 : 0,
-                isAmrap: attrs.contains("amrap") ? 1 : 0,
-                side: self.stringField(record, "side"),
-                updatedAt: now
-            )
-            if existing != nil { try row.update(db) } else { try row.insert(db) }
-
-            // Replace measurements from CK record fields (dual-read: old-format CKRecords)
-            try db.execute(
-                sql: "DELETE FROM set_measurements WHERE set_id = ? AND parent_type = 'session'",
-                arguments: [setId]
-            )
-            try self.insertMeasurementsFromCKRecord(
-                record, setId: setId, parentType: "session", role: "target", now: now, in: db
-            )
-            try self.insertMeasurementsFromCKRecord(
-                record, setId: setId, parentType: "session", role: "actual", now: now, in: db
-            )
-
-            return true
-        }
-    }
-
-    private func mergeUserSettings(_ record: CKRecord, dbQueue: DatabaseQueue) throws -> Bool {
-        let remoteUpdatedAt = dateField(record, "updatedAt")
-        return try dbQueue.write { db in
-            let existing = try UserSettingsRow.fetchOne(db)
-            if let existing, !self.remoteIsNewer(remoteDate: remoteUpdatedAt, localUpdatedAt: existing.updatedAt) {
-                return false
-            }
-            if let existing {
-                let updatedAt = self.dateToISO(remoteUpdatedAt) ?? existing.updatedAt
-                let row = UserSettingsRow(
-                    id: existing.id,
-                    defaultWeightUnit: self.stringField(record, "defaultWeightUnit") ?? existing.defaultWeightUnit,
-                    enableWorkoutTimer: Int(self.int64Field(record, "enableWorkoutTimer")
-                        ?? Int64(existing.enableWorkoutTimer)),
-                    autoStartRestTimer: Int(self.int64Field(record, "autoStartRestTimer")
-                        ?? Int64(existing.autoStartRestTimer)),
-                    theme: self.stringField(record, "theme") ?? existing.theme,
-                    notificationsEnabled: Int(self.int64Field(record, "notificationsEnabled")
-                        ?? Int64(existing.notificationsEnabled)),
-                    customPromptAddition: self.stringField(record, "customPromptAddition")
-                        ?? existing.customPromptAddition,
-                    anthropicApiKeyStatus: existing.anthropicApiKeyStatus, // Never sync
-                    healthkitEnabled: Int(self.int64Field(record, "healthKitEnabled")
-                        ?? Int64(existing.healthkitEnabled)),
-                    liveActivitiesEnabled: Int(self.int64Field(record, "liveActivitiesEnabled")
-                        ?? Int64(existing.liveActivitiesEnabled)),
-                    keepScreenAwake: Int(self.int64Field(record, "keepScreenAwake") ?? Int64(existing.keepScreenAwake)),
-                    showOpenInClaudeButton: Int(self.int64Field(record, "showOpenInClaudeButton")
-                        ?? Int64(existing.showOpenInClaudeButton)),
-                    developerModeEnabled: existing.developerModeEnabled,
-                    countdownSoundsEnabled: Int(self.int64Field(record, "countdownSoundsEnabled")
-                        ?? Int64(existing.countdownSoundsEnabled)),
-                    hasAcceptedDisclaimer: existing.hasAcceptedDisclaimer, // Never sync — local-only
-                    defaultTimerCountdown: Int(self.int64Field(record, "defaultTimerCountdown")
-                        ?? Int64(existing.defaultTimerCountdown)),
-                    defaultWeightStepLbs: self.doubleField(record, "defaultWeightStepLbs")
-                        ?? existing.defaultWeightStepLbs,
-                    aiPromptIncludeFormatPointer: Int(self.int64Field(record, "aiPromptIncludeFormatPointer")
-                        ?? Int64(existing.aiPromptIncludeFormatPointer)),
-                    aiPromptIncludeRecentWorkouts: Int(self.int64Field(record, "aiPromptIncludeRecentWorkouts")
-                        ?? Int64(existing.aiPromptIncludeRecentWorkouts)),
-                    aiPromptIncludeProgression: Int(self.int64Field(record, "aiPromptIncludeProgression")
-                        ?? Int64(existing.aiPromptIncludeProgression)),
-                    aiPromptIncludeEquipment: Int(self.int64Field(record, "aiPromptIncludeEquipment")
-                        ?? Int64(existing.aiPromptIncludeEquipment)),
-                    homeTiles: self.stringField(record, "homeTiles") ?? existing.homeTiles,
-                    createdAt: existing.createdAt,
-                    updatedAt: updatedAt
-                )
-                try row.update(db)
-                return true
-            } else {
-                let now = self.isoFormatter.string(from: Date())
-                let row = UserSettingsRow(
-                    id: IDGenerator.generate(),
-                    defaultWeightUnit: self.stringField(record, "defaultWeightUnit") ?? "lbs",
-                    enableWorkoutTimer: Int(self.int64Field(record, "enableWorkoutTimer") ?? 1),
-                    autoStartRestTimer: Int(self.int64Field(record, "autoStartRestTimer") ?? 1),
-                    theme: self.stringField(record, "theme") ?? "auto",
-                    notificationsEnabled: Int(self.int64Field(record, "notificationsEnabled") ?? 1),
-                    customPromptAddition: self.stringField(record, "customPromptAddition"),
-                    anthropicApiKeyStatus: "not_set",
-                    healthkitEnabled: Int(self.int64Field(record, "healthKitEnabled") ?? 0),
-                    liveActivitiesEnabled: Int(self.int64Field(record, "liveActivitiesEnabled") ?? 1),
-                    keepScreenAwake: Int(self.int64Field(record, "keepScreenAwake") ?? 1),
-                    showOpenInClaudeButton: Int(self.int64Field(record, "showOpenInClaudeButton") ?? 0),
-                    developerModeEnabled: 0,
-                    countdownSoundsEnabled: Int(self.int64Field(record, "countdownSoundsEnabled") ?? 1),
-                    hasAcceptedDisclaimer: 0, // New device — must accept again
-                    defaultTimerCountdown: Int(self.int64Field(record, "defaultTimerCountdown") ?? 0),
-                    defaultWeightStepLbs: self.doubleField(record, "defaultWeightStepLbs") ?? 2.5,
-                    aiPromptIncludeFormatPointer: Int(self.int64Field(record, "aiPromptIncludeFormatPointer") ?? 1),
-                    aiPromptIncludeRecentWorkouts: Int(self.int64Field(record, "aiPromptIncludeRecentWorkouts") ?? 1),
-                    aiPromptIncludeProgression: Int(self.int64Field(record, "aiPromptIncludeProgression") ?? 1),
-                    aiPromptIncludeEquipment: Int(self.int64Field(record, "aiPromptIncludeEquipment") ?? 1),
-                    homeTiles: self.stringField(record, "homeTiles"),
-                    createdAt: now,
-                    updatedAt: self.dateToISO(remoteUpdatedAt) ?? now
-                )
-                try row.insert(db)
-                return true
-            }
-        }
+    /// Shared fallback for a CKRecord whose type no merge route recognizes.
+    func logUnknownRecordType(_ record: CKRecord) -> Bool {
+        Logger.shared.warn(.sync, "Unknown record type for merge: \(record.recordType)")
+        return false
     }
 
     // MARK: - Record Lookup
@@ -710,49 +216,11 @@ final class CKRecordMapper {
             // Build the fresh record inside the read, then rehydrate its change tag OUTSIDE
             // the read (applyStoredSystemFields opens its own read → would be reentrant here).
             let fresh = try dbQueue.read { db -> CKRecord? in
-                if let gym = try GymRow.fetchOne(db, key: id) {
-                    if gym.deletedAt != nil { return nil }
-                    return self.toCKRecord(gym, zoneID: zoneID)
-                }
-                if let eq = try GymEquipmentRow.fetchOne(db, key: id) {
-                    if eq.deletedAt != nil { return nil }
-                    return self.toCKRecord(eq, zoneID: zoneID)
-                }
-                if let plan = try WorkoutPlanRow.fetchOne(db, key: id) {
-                    return self.toCKRecord(plan, zoneID: zoneID)
-                }
-                if let ex = try PlannedExerciseRow.fetchOne(db, key: id) {
-                    return self.toCKRecord(ex, zoneID: zoneID)
-                }
-                if let ps = try PlannedSetRow.fetchOne(db, key: id) {
-                    let measurements = try SetMeasurementRow
-                        .filter(Column("set_id") == id)
-                        .filter(Column("parent_type") == "planned")
-                        .fetchAll(db)
-                    return self.toCKRecord(ps, measurements: measurements, zoneID: zoneID)
-                }
-                if let session = try WorkoutSessionRow.fetchOne(db, key: id) {
-                    return self.toCKRecord(session, zoneID: zoneID)
-                }
-                if let se = try SessionExerciseRow.fetchOne(db, key: id) {
-                    return self.toCKRecord(se, zoneID: zoneID)
-                }
-                if let ss = try SessionSetRow.fetchOne(db, key: id) {
-                    let measurements = try SetMeasurementRow
-                        .filter(Column("set_id") == id)
-                        .filter(Column("parent_type") == "session")
-                        .fetchAll(db)
-                    return self.toCKRecord(ss, measurements: measurements, zoneID: zoneID)
-                }
-                if let settings = try UserSettingsRow.fetchOne(db) {
-                    if settings.id == id || id == "user-settings" {
-                        return self.toCKRecord(settings, zoneID: zoneID)
-                    }
-                }
-                if let measurement = try SetMeasurementRow.fetchOne(db, key: id) {
-                    return self.toCKRecord(measurement, zoneID: zoneID)
-                }
-                return nil
+                try self.freshGymTierRecord(id: id, zoneID: zoneID, db: db)
+                    ?? self.freshPlanTierRecord(id: id, zoneID: zoneID, db: db)
+                    ?? self.freshSessionTierRecord(id: id, zoneID: zoneID, db: db)
+                    ?? self.freshSettingsRecord(id: id, zoneID: zoneID, db: db)
+                    ?? self.freshMeasurementRecord(id: id, zoneID: zoneID, db: db)
             }
             guard let fresh else { return nil }
             return applyStoredSystemFields(to: fresh)
@@ -784,93 +252,6 @@ final class CKRecordMapper {
                 try db.execute(sql: "DELETE FROM \(table) WHERE id = ?", arguments: [id])
                 if db.changesCount > 0 { return }
             }
-        }
-    }
-
-    // MARK: - Active Session Protection
-
-    /// IDs belonging to an in-progress workout session that must not be deleted or overwritten by sync.
-    struct ActiveSessionProtectedIds {
-        let sessionId: String?
-        let exerciseIds: Set<String>
-        let setIds: Set<String>
-        let planId: String?
-        let plannedExerciseIds: Set<String>
-        let plannedSetIds: Set<String>
-
-        /// All protected IDs keyed by CloudKit record type.
-        var byRecordType: [String: Set<String>] {
-            var map: [String: Set<String>] = [:]
-            if let sid = sessionId { map["WorkoutSession"] = [sid] }
-            if !exerciseIds.isEmpty { map["SessionExercise"] = exerciseIds }
-            if !setIds.isEmpty { map["SessionSet"] = setIds }
-            if let pid = planId { map["WorkoutPlan"] = [pid] }
-            if !plannedExerciseIds.isEmpty { map["PlannedExercise"] = plannedExerciseIds }
-            if !plannedSetIds.isEmpty { map["PlannedSet"] = plannedSetIds }
-            return map
-        }
-
-        static let empty = ActiveSessionProtectedIds(
-            sessionId: nil, exerciseIds: [], setIds: [],
-            planId: nil, plannedExerciseIds: [], plannedSetIds: []
-        )
-    }
-
-    /// Query the database for the active (in_progress) session and collect all IDs that belong to it,
-    /// including the parent WorkoutPlan's PlannedExercise and PlannedSet records.
-    func getActiveSessionProtectedIds() -> ActiveSessionProtectedIds {
-        do {
-            let dbQueue = try dbManager.database()
-            return try dbQueue.read { db in
-                guard let sessionRow = try Row.fetchOne(
-                    db,
-                    sql: "SELECT id, workout_template_id FROM workout_sessions WHERE status = 'in_progress' LIMIT 1"
-                ),
-                      let sessionId: String = sessionRow["id"] else {
-                    return .empty
-                }
-
-                let exSql = "SELECT id FROM session_exercises WHERE workout_session_id = ?"
-                let exerciseRows = try Row.fetchAll(db, sql: exSql, arguments: [sessionId])
-                let exerciseIds = Set(exerciseRows.compactMap { $0["id"] as String? })
-
-                var setIds = Set<String>()
-                if !exerciseIds.isEmpty {
-                    let placeholders = exerciseIds.map { _ in "?" }.joined(separator: ",")
-                    let setSql = "SELECT id FROM session_sets WHERE session_exercise_id IN (\(placeholders))"
-                    let setRows = try Row.fetchAll(db, sql: setSql, arguments: StatementArguments(Array(exerciseIds)))
-                    setIds = Set(setRows.compactMap { $0["id"] as String? })
-                }
-
-                let planId: String? = sessionRow["workout_template_id"]
-                var plannedExerciseIds = Set<String>()
-                var plannedSetIds = Set<String>()
-
-                if let planId, !planId.isEmpty {
-                    let peSql = "SELECT id FROM template_exercises WHERE workout_template_id = ?"
-                    let peRows = try Row.fetchAll(db, sql: peSql, arguments: [planId])
-                    plannedExerciseIds = Set(peRows.compactMap { $0["id"] as String? })
-
-                    if !plannedExerciseIds.isEmpty {
-                        let pePlaceholders = plannedExerciseIds.map { _ in "?" }.joined(separator: ",")
-                        let psRows = try Row.fetchAll(
-                            db,
-                            sql: "SELECT id FROM template_sets "
-                                + "WHERE template_exercise_id IN (\(pePlaceholders))",
-                            arguments: StatementArguments(Array(plannedExerciseIds))
-                        )
-                        plannedSetIds = Set(psRows.compactMap { $0["id"] as String? })
-                    }
-                }
-
-                return ActiveSessionProtectedIds(
-                    sessionId: sessionId, exerciseIds: exerciseIds, setIds: setIds,
-                    planId: planId, plannedExerciseIds: plannedExerciseIds, plannedSetIds: plannedSetIds
-                )
-            }
-        } catch {
-            Logger.shared.error(.app, "Failed to query active session for sync protection", error: error)
-            return .empty
         }
     }
 }
