+++ mobile-apps/ios/LiftMark/Services/CKSyncConflictResolver.swift
@@ -69,98 +69,9 @@ final class CKSyncConflictResolver: @unchecked Sendable {
 
         // Handle failures
         for failedSave in event.failedRecordSaves {
-            let record = failedSave.record
-            let recordName = record.recordID.recordName
-            let recordType = record.recordType
-            let error = failedSave.error
-
-            switch error.code {
-            case .serverRecordChanged:
+            let wasConflict = handleFailedSave(failedSave, removePendingType: removePendingType, engine: engine)
+            if wasConflict {
                 conflicts += 1
-                handleServerRecordChanged(recordName: recordName, recordType: recordType, error: error)
-
-            case .networkFailure, .networkUnavailable:
-                Logger.shared.warn(.sync, "[sync-engine] Network unavailable for \(recordType)/\(recordName) " +
-                    "(CKError \(error.code.rawValue)), CKSyncEngine will retry")
-                // Transient — breadcrumb only, don't capture.
-                CrashReporter.shared.addBreadcrumb("sync.networkRetry", category: .sync,
-                    metadata: ["recordType": recordType, "errorCode": "\(error.code.rawValue)"])
-
-            case .quotaExceeded:
-                Logger.shared.error(.sync, "[sync-engine] iCloud quota exceeded for \(recordType)/\(recordName) " +
-                    "(CKError \(error.code.rawValue))")
-                CrashReporter.shared.captureError(error, category: .sync, metadata: [
-                    "recordType": recordType, "errorCode": "\(error.code.rawValue)", "errorDomain": CKErrorDomain
-                ])
-
-            case .notAuthenticated:
-                Logger.shared.error(.sync, "[sync-engine] Not authenticated for \(recordType)/\(recordName) " +
-                    "(CKError \(error.code.rawValue))")
-                CrashReporter.shared.captureError(error, category: .sync, metadata: [
-                    "recordType": recordType, "errorCode": "\(error.code.rawValue)", "errorDomain": CKErrorDomain
-                ])
-
-            case .unknownItem:
-                Logger.shared.warn(.sync, "[sync-engine] Unknown item \(recordType)/\(recordName) " +
-                    "(CKError \(error.code.rawValue), parent likely deleted), removing from pending")
-                let ckRecordID = record.recordID
-                engine?.state.remove(pendingRecordZoneChanges: [.saveRecord(ckRecordID)])
-                removePendingType(recordName)
-
-            case .partialFailure:
-                handlePartialFailure(recordName: recordName, recordType: recordType, error: error)
-
-            case .invalidArguments, .serverRejectedRequest:
-                // CKError 12 (invalidArguments) and 2006 (serverRejectedRequest) both
-                // indicate the local record doesn't match the deployed CloudKit container
-                // schema — typically a field present locally but not in Production, or
-                // a type mismatch. Capture the field names so we can identify which
-                // schema element is missing without needing to repro.
-                let fields = record.allKeys().sorted().joined(separator: ",")
-                Logger.shared.error(
-                    .sync,
-                    "[sync-engine] Server rejected \(recordType)/\(recordName) "
-                        + "(CKError \(error.code.rawValue) "
-                        + "(\(Self.errorCodeName(error.code)))) "
-                        + "— likely schema drift. Fields sent: [\(fields)]"
-                )
-                let metadata: [String: String] = [
-                    "recordType": recordType,
-                    "recordFields": fields,
-                    "errorCode": "\(error.code.rawValue)",
-                    "errorDomain": CKErrorDomain,
-                    "tag": "schema-drift-suspected"
-                ]
-                CrashReporter.shared.captureErrorOncePerBuild(
-                    key: "ck-reject-\(error.code.rawValue)-\(recordType)",
-                    error: error,
-                    breadcrumb: "sync.schemaDrift.repeat",
-                    category: .sync,
-                    metadata: metadata
-                )
-
-            default:
-                Logger.shared.error(
-                    .sync,
-                    "[sync-engine] Failed to save \(recordType)/\(recordName): "
-                        + "CKError \(error.code.rawValue) "
-                        + "(\(Self.errorCodeName(error.code))) "
-                        + "— \(error.localizedDescription)"
-                )
-                let fields = record.allKeys().sorted().joined(separator: ",")
-                let metadata: [String: String] = [
-                    "recordType": recordType,
-                    "recordFields": fields,
-                    "errorCode": "\(error.code.rawValue)",
-                    "errorDomain": CKErrorDomain
-                ]
-                CrashReporter.shared.captureErrorOncePerBuild(
-                    key: "ck-fail-\(error.code.rawValue)-\(recordType)",
-                    error: error,
-                    breadcrumb: "sync.saveFailed.repeat",
-                    category: .sync,
-                    metadata: metadata
-                )
             }
         }
 
@@ -168,6 +79,119 @@ final class CKSyncConflictResolver: @unchecked Sendable {
     }
 
     // MARK: - Private
+
+    /// Dispatch a single failed record save by CKError code. Returns `true` when the
+    /// failure was a `serverRecordChanged` conflict (so the caller can count it).
+    private func handleFailedSave(
+        _ failedSave: CKSyncEngine.Event.SentRecordZoneChanges.FailedRecordSave,
+        removePendingType: (String) -> Void,
+        engine: CKSyncEngine?
+    ) -> Bool {
+        let record = failedSave.record
+        let recordName = record.recordID.recordName
+        let recordType = record.recordType
+        let error = failedSave.error
+
+        switch error.code {
+        case .serverRecordChanged:
+            handleServerRecordChanged(recordName: recordName, recordType: recordType, error: error)
+            return true
+
+        case .networkFailure, .networkUnavailable:
+            Logger.shared.warn(.sync, "[sync-engine] Network unavailable for \(recordType)/\(recordName) " +
+                "(CKError \(error.code.rawValue)), CKSyncEngine will retry")
+            // Transient — breadcrumb only, don't capture.
+            CrashReporter.shared.addBreadcrumb("sync.networkRetry", category: .sync,
+                metadata: ["recordType": recordType, "errorCode": "\(error.code.rawValue)"])
+
+        case .quotaExceeded:
+            Logger.shared.error(.sync, "[sync-engine] iCloud quota exceeded for \(recordType)/\(recordName) " +
+                "(CKError \(error.code.rawValue))")
+            CrashReporter.shared.captureError(error, category: .sync, metadata: [
+                "recordType": recordType, "errorCode": "\(error.code.rawValue)", "errorDomain": CKErrorDomain
+            ])
+
+        case .notAuthenticated:
+            Logger.shared.error(.sync, "[sync-engine] Not authenticated for \(recordType)/\(recordName) " +
+                "(CKError \(error.code.rawValue))")
+            CrashReporter.shared.captureError(error, category: .sync, metadata: [
+                "recordType": recordType, "errorCode": "\(error.code.rawValue)", "errorDomain": CKErrorDomain
+            ])
+
+        case .unknownItem:
+            Logger.shared.warn(.sync, "[sync-engine] Unknown item \(recordType)/\(recordName) " +
+                "(CKError \(error.code.rawValue), parent likely deleted), removing from pending")
+            let ckRecordID = record.recordID
+            engine?.state.remove(pendingRecordZoneChanges: [.saveRecord(ckRecordID)])
+            removePendingType(recordName)
+
+        case .partialFailure:
+            handlePartialFailure(recordName: recordName, recordType: recordType, error: error)
+
+        case .invalidArguments, .serverRejectedRequest:
+            handleSchemaDriftRejection(record: record, recordName: recordName, recordType: recordType, error: error)
+
+        default:
+            handleUnexpectedSaveFailure(record: record, recordName: recordName, recordType: recordType, error: error)
+        }
+        return false
+    }
+
+    /// CKError 12 (invalidArguments) and 2006 (serverRejectedRequest) both
+    /// indicate the local record doesn't match the deployed CloudKit container
+    /// schema — typically a field present locally but not in Production, or
+    /// a type mismatch. Capture the field names so we can identify which
+    /// schema element is missing without needing to repro.
+    private func handleSchemaDriftRejection(record: CKRecord, recordName: String, recordType: String, error: CKError) {
+        let fields = record.allKeys().sorted().joined(separator: ",")
+        Logger.shared.error(
+            .sync,
+            "[sync-engine] Server rejected \(recordType)/\(recordName) "
+                + "(CKError \(error.code.rawValue) "
+                + "(\(Self.errorCodeName(error.code)))) "
+                + "— likely schema drift. Fields sent: [\(fields)]"
+        )
+        let metadata: [String: String] = [
+            "recordType": recordType,
+            "recordFields": fields,
+            "errorCode": "\(error.code.rawValue)",
+            "errorDomain": CKErrorDomain,
+            "tag": "schema-drift-suspected"
+        ]
+        CrashReporter.shared.captureErrorOncePerBuild(
+            key: "ck-reject-\(error.code.rawValue)-\(recordType)",
+            error: error,
+            breadcrumb: "sync.schemaDrift.repeat",
+            category: .sync,
+            metadata: metadata
+        )
+    }
+
+    /// Catch-all for CKError codes without a dedicated branch — logs and captures with the
+    /// same `recordFields` diagnostic payload as the schema-drift path.
+    private func handleUnexpectedSaveFailure(record: CKRecord, recordName: String, recordType: String, error: CKError) {
+        Logger.shared.error(
+            .sync,
+            "[sync-engine] Failed to save \(recordType)/\(recordName): "
+                + "CKError \(error.code.rawValue) "
+                + "(\(Self.errorCodeName(error.code))) "
+                + "— \(error.localizedDescription)"
+        )
+        let fields = record.allKeys().sorted().joined(separator: ",")
+        let metadata: [String: String] = [
+            "recordType": recordType,
+            "recordFields": fields,
+            "errorCode": "\(error.code.rawValue)",
+            "errorDomain": CKErrorDomain
+        ]
+        CrashReporter.shared.captureErrorOncePerBuild(
+            key: "ck-fail-\(error.code.rawValue)-\(recordType)",
+            error: error,
+            breadcrumb: "sync.saveFailed.repeat",
+            category: .sync,
+            metadata: metadata
+        )
+    }
 
     private func handleServerRecordChanged(recordName: String, recordType: String, error: CKError) {
         if let serverRecord = error.serverRecord {
